change: Moved some of the types around and implemented grouping for the classes
