+++ prerequisite_parser/parsed_types.py
@@ -1,4 +1,3 @@
-from enum import Enum, unique
 import json
 
 from bs4 import NavigableString
@@ -107,97 +106,3 @@ def extract_classes(input: str) -> List[str]:
     if "Frosh" in input:
         levels.append("Frosh")
     return levels
-
-
-@unique
-class Operator(Enum):
-    Or = 1
-    And = 2
-
-    @staticmethod
-    def parse(text: str) -> Self:
-        if text == "or":
-            return Operator.Or
-        elif text == "and":
-            return Operator.And
-        raise Exception(f'Unknown operator type: "{text}"')
-
-    def __str__(self) -> str:
-        return self.name.lower()
-
-
-class ParsedPrerequisite(json.JSONEncoder):
-    prefixed_operator: Operator | None = None
-
-
-class SingleClass(ParsedPrerequisite):
-    class_name: str
-
-    def __init__(
-        self, class_name: str, prefixed_operator: Operator | None = None
-    ) -> None:
-        self.class_name = class_name
-        self.prefixed_operator = prefixed_operator
-
-    def to_json(self):
-        return self.__dict__()
-
-    def __repr__(self) -> str:
-        return f"SingleClass {{ prefixed_operator: {self.prefixed_operator}, class_name: {self.class_name} }}"
-
-    def __dict__(self) -> Dict[str, Any]:
-        return {"course": self.class_name, "type": "course"}
-
-
-class ClassGroup(ParsedPrerequisite):
-    nested_classes: List[ParsedPrerequisite]
-
-    @staticmethod
-    def from_prereq_list(prereqs: List[ParsedPrerequisite]) -> Self:
-        # The grouping type is everything after the first item
-        grouping_type = prereqs[1].prefixed_operator
-        return ClassGroup(
-            nested_classes=prereqs,
-            prefixed_operator=grouping_type,
-        )
-
-    def __init__(
-        self,
-        nested_classes: List[ParsedPrerequisite],
-        prefixed_operator: Operator | None = None,
-    ):
-        self.nested_classes = nested_classes
-        self.prefixed_operator = prefixed_operator
-
-    def to_json(self):
-        return json.dumps(self.__dict__(), indent=2)
-
-    def __repr__(self) -> str:
-        return f"ClassGroup {{ type: {self.prefixed_operator}, nested: {self.nested_classes} }}"
-
-    def __dict__(self) -> Dict[str, Any]:
-        return {
-            "nested": list(map(lambda item: item.__dict__(), self.nested_classes)),
-            "type": str(self.prefixed_operator),
-        }
-
-
-class Prerequisite:
-    @staticmethod
-    def from_parsed_prereq_list(prereqs: List[ParsedPrerequisite]) -> Self:
-        print(prereqs)
-
-
-class PrereqCourse(Prerequisite):
-    course_name: str
-
-    def __dict__(self) -> Dict[str, Any]:
-        return {"course": self.course_name, "type": "course"}
-
-
-class PrereqGroup(Prerequisite):
-    nested_courses: List[Prerequisite]
-    group_operator: Operator
-
-    def __dict__(self) -> Dict[str, Any]:
-        return {"nested": self.nested_courses, "type": self.group_operator}
